add browserify and grunt and loosely add files
