+++ package.json
@@ -17,6 +17,13 @@
     "url": "git@github.com:mimiflynn/console-game.git"
   },
   "devDependencies": {
-    "http-server": "^0.8.0"
+    "browserify": "^11.0.1",
+    "http-server": "^0.8.0",
+    "grunt": "^0.4.5",
+    "grunt-browserify": "^4.0.0",
+    "grunt-contrib-clean": "^0.6.0",
+    "grunt-contrib-jshint": "^0.11.2",
+    "grunt-contrib-watch": "^0.6.1"
+    
   }
 }
