+++ public/index.html
@@ -8,7 +8,6 @@
 	</head>
 	<body>
 		<h1>open that console, yo!</h1>
-		<script type="application/javascript" src="js/The-Mansion.js"></script>
-		<script type="application/javascript" src="js/complex-game.js"></script>
+		<script type="application/javascript" src="js/console-game.js"></script>
 	</body>
 </html>
