+++ source/app.js
@@ -1,0 +1,1 @@
+var TheMansion = require('./the-mansion/game');
